clean up: split javascript
